Prepare for refactoring
