+++ lib/budget/supportedCurrencies.js
