+++ lib/test/unit/budget/budgetFormatter.spec.js
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 
-const { prepareValue } = require('../../budgetFormatter');
+const { prepareValue } = require('../../../budget/budgetFormatter');
 
 describe('budgetFormatter plugin: prepareValue method', () => {
   test('Should return the original input if non-string value is passed', () => {
