+++ lib/test/unit/budget/extractBudgetData.spec.js
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 
-const extractBudgetData = require('../../extractBudgetData');
+const extractBudgetData = require('../../../budget/extractBudgetData');
 
 describe('extractBudgetData helper', () => {
   test('Should return an empty object when no arguments', () => {
