+++ services/ingestion/etl/agri/csv/src/lib/transform.js
@@ -1,6 +1,6 @@
 // @flow
 
-import sanitizeBudgetItem from '@eubfr/lib/budgetFormatter';
+import sanitizeBudgetItem from '@eubfr/lib/budget/budgetFormatter';
 import getCountryCode from '@eubfr/lib/getCountryCode';
 import type { Project } from '@eubfr/types';
