+++ services/ingestion/etl/inforegio/xml/src/lib/transform.js
@@ -1,7 +1,7 @@
 // @flow
 
-import sanitizeBudgetItem from '@eubfr/lib/budgetFormatter';
-import extractBudgetData from '@eubfr/lib/extractBudgetData';
+import sanitizeBudgetItem from '@eubfr/lib/budget/budgetFormatter';
+import extractBudgetData from '@eubfr/lib/budget/extractBudgetData';
 import getCountryCode from '@eubfr/lib/getCountryCode';
 import type { Project } from '@eubfr/types';
